Working Build
This is the first working build! This version successfully implements the Google Task API. The user is first instructed to log into google, and then is able to add Canvas scraped assignments into any of your task lists. To ensure usability when tokens expire,  the extension automatically relogs you in with a new token.